requirements.txt
boto3>=1.26.0
PyYAML>=6.0
tqdm>=4.64.0
